refactor(color): split gray.rs → gray/{mod,kernels}.rs; move bgr_from_rgb to rgb.rs
Follow the filter/warp/resize/morphology convention:
- gray/mod.rs  — public Image-typed API (gray_from_rgb, gray_from_rgb_{u8,f32}, rgb_from_gray)
- gray/kernels.rs — constants, par_strip_dispatch, slice dispatchers,
                   and all NEON/AVX2/scalar kernel implementations
- bgr_from_rgb was misplaced in gray.rs; moved to rgb.rs with its test

No behaviour change; all 13 gray + rgb tests pass.
